+++ packages/backan/CHANGELOG.md
@@ -1,5 +1,14 @@
 # backan
 
+## 0.0.11
+
+### Patch Changes
+
+- Update create package, update readmes
+
+- Updated dependencies []:
+  - @backan/core@0.0.11
+
 ## 0.0.6
 
 ### Patch Changes