+++ packages/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backan/docs
 
+## 0.0.11
+
+### Patch Changes
+
+- Update create package, update readmes
+
 ## 0.0.6
 
 ### Patch Changes